http://jira.bytatech.com:8080/browse/IN-8845

+++ src/main/java/com/lxisoft/service/UserExtraService.java
@@ -43,7 +43,7 @@ public class UserExtraService {
 		Optional<User> optUser=userRepo.findOneByLogin(login);
 		User currentUser=optUser.get();
 		log.debug("user currently logged is :"+currentUser);
-		Optional<UserExtra> optExtra=extraRepo.findById((long) currentUser.getId());
+		Optional<UserExtra> optExtra=extraRepo.findById( currentUser.getId());
 		UserExtra userExtra=optExtra.get();
 		// userExtra will already be saved while creating new user. so currentUser -> userExtra obtained
 		return userExtra;
@@ -58,7 +58,7 @@ public class UserExtraService {
 			long id=Integer.parseInt(uid);
 			Optional<User>user=userRepo.findById(id);
 			User currentUser=user.get();
-			Optional<UserExtra> optExtra=extraRepo.findById((long) currentUser.getId());
+			Optional<UserExtra> optExtra=extraRepo.findById(currentUser.getId());
 			UserExtra userExtra=optExtra.get();
 			
 		
@@ -66,7 +66,6 @@ public class UserExtraService {
 	}
 
 	public void save(@Valid User user) {
-//		String pass = user.getPassword();
 		BCryptPasswordEncoder encode=new BCryptPasswordEncoder(); 
 		user.setPassword(encode.encode(user.getPassword()));
 		userRepo.save(user);
